TrackGeneratorStrategy: abstract track layers.
Take the for(...) f() functions which generate tracks and make them
into objects which maintain the state of the for loop. This will allow
us to generate tracks more continuously, seems to reduce the number of
entities created, and may allow us to limit the number of planks we
create.