Show "PRESS R" when the player loses.
Re-use the player shader program again. Really need to rename that...

Make the player texture program have a color that can be passed in to
modify the color drawn on screen. This allows us to fade in our PRESS R
dialogue.

Add a StopWatch class to timer.h which allows this dialogue to keep
track of how long ago the player actually lost.